Move task to tasks.py
